requirements.txt
# API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# HTTP Client
httpx==0.26.0

# Document Processing
python-docx==1.1.0
PyMuPDF==1.23.8
Pillow==10.2.0

# Database
# SQLite is built-in

# Data
numpy==1.26.3
pydantic==2.5.3

# Utils
python-dotenv==1.0.0

# Security
cryptography>=41.0.0

# Optional: Local LLM
# ollama (install separately)

# Code Execution - E2B (Cloud - Recommended)
e2b-code-interpreter>=1.0.0

# Code Execution - Local (Docker) - optional, requires Docker Desktop
# llm-sandbox[docker]>=0.5.0
# docker>=6.0.0

# Document Generation
reportlab>=4.0.0
